pyproject.toml
[project]
name = "macro-finance-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fredapi>=0.5.2",
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.0",
    "numpy>=2.2.2",
    "polars>=1.22.0",
    "pyarrow>=19.0.0",
    "python-dotenv>=1.0.1",
    "seaborn>=0.13.2",
    "statsmodels>=0.14.4",
    "yfinance>=0.2.52",
]
